incorporate fix for Go 1.16 from upstream PR #138 on top of v0.19.8
As reported in https://github.com/go-openapi/validate/issues/137, changes in Go 1.16's `reflect` package cause the type validator to fail now. Incorporate a fix from https://github.com/go-openapi/validate/pull/138 to fix that.

For the time being, for internal compatibility and transitive dependency issues, we need to stick with v0.19.8. But to unblock upgrading to Go 1.16.0, we need this hotfix.